Fetch Author Name from Authors in JSON Response
Instead of relying on `by_statement` (which seems to require
`jscmd=details`), the more stable `authors` key (which is available in a
`jscmd=data` request) is searched, and the first author in the list is
found.

This can be further updated to handle multiple authors if a book has them.